+++ usr.sbin/bgpd/session.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: session.h,v 1.131 2019/02/18 09:58:19 claudio Exp $ */
+/*	$OpenBSD: session.h,v 1.132 2019/02/19 09:13:23 claudio Exp $ */
 
 /*
  * Copyright (c) 2003, 2004 Henning Brauer <henning@openbsd.org>
@@ -149,22 +149,22 @@ struct ctl_conn {
 TAILQ_HEAD(ctl_conns, ctl_conn)	ctl_conns;
 
 struct peer_stats {
-	u_int64_t		 msg_rcvd_open;
-	u_int64_t		 msg_rcvd_update;
-	u_int64_t		 msg_rcvd_notification;
-	u_int64_t		 msg_rcvd_keepalive;
-	u_int64_t		 msg_rcvd_rrefresh;
-	u_int64_t		 msg_sent_open;
-	u_int64_t		 msg_sent_update;
-	u_int64_t		 msg_sent_notification;
-	u_int64_t		 msg_sent_keepalive;
-	u_int64_t		 msg_sent_rrefresh;
-	u_int64_t		 prefix_rcvd_update;
-	u_int64_t		 prefix_rcvd_withdraw;
-	u_int64_t		 prefix_rcvd_eor;
-	u_int64_t		 prefix_sent_update;
-	u_int64_t		 prefix_sent_withdraw;
-	u_int64_t		 prefix_sent_eor;
+	unsigned long long	 msg_rcvd_open;
+	unsigned long long	 msg_rcvd_update;
+	unsigned long long	 msg_rcvd_notification;
+	unsigned long long	 msg_rcvd_keepalive;
+	unsigned long long	 msg_rcvd_rrefresh;
+	unsigned long long	 msg_sent_open;
+	unsigned long long	 msg_sent_update;
+	unsigned long long	 msg_sent_notification;
+	unsigned long long	 msg_sent_keepalive;
+	unsigned long long	 msg_sent_rrefresh;
+	unsigned long long	 prefix_rcvd_update;
+	unsigned long long	 prefix_rcvd_withdraw;
+	unsigned long long	 prefix_rcvd_eor;
+	unsigned long long	 prefix_sent_update;
+	unsigned long long	 prefix_sent_withdraw;
+	unsigned long long	 prefix_sent_eor;
 	time_t			 last_updown;
 	time_t			 last_read;
 	u_int32_t		 prefix_cnt;
